Move utility functions from viewholders to presenters
