+++ app/src/main/java/greenberg/moviedbshell/MosbyImpl/SearchPresenter.kt
@@ -2,11 +2,17 @@ package greenberg.moviedbshell.MosbyImpl
 
 import android.support.v7.widget.LinearLayoutManager
 import android.support.v7.widget.RecyclerView
+import android.widget.ImageView
+import com.bumptech.glide.Glide
+import com.bumptech.glide.request.RequestOptions
 import com.hannesdorfmann.mosby3.mvp.MvpBasePresenter
 import greenberg.moviedbshell.Models.SearchModels.SearchResultsItem
+import greenberg.moviedbshell.R
 import greenberg.moviedbshell.RetrofitHelpers.RetrofitHelper
+import greenberg.moviedbshell.ViewHolders.SearchResultsAdapter
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.schedulers.Schedulers
+import java.text.SimpleDateFormat
 
 class SearchPresenter : MvpBasePresenter<ZephyrrSearchView>() {
 
@@ -84,5 +90,38 @@ class SearchPresenter : MvpBasePresenter<ZephyrrSearchView>() {
                             view.showError(throwable, false)
                     }
                 })
+    }
+
+    fun fetchPoster(cardItemPosterView: ImageView, item: SearchResultsItem) {
+        //Load poster art
+        when (item.mediaType) {
+            SearchResultsAdapter.MEDIA_TYPE_MOVIE, SearchResultsAdapter.MEDIA_TYPE_TV -> {
+                item.posterPath
+            }
+            SearchResultsAdapter.MEDIA_TYPE_PERSON -> {
+                item.profilePath
+            }
+            else -> {
+                //Don't fetch poster if there is no poster art
+                return
+            }
+        }?.let {
+            Glide.with(cardItemPosterView)
+                    //TODO: potentially hacky way to get context
+                    .load(cardItemPosterView.context.getString(R.string.poster_url_substitution, it))
+                    .apply { RequestOptions().centerCrop() }
+                    .into(cardItemPosterView)
+        }
+    }
+
+    fun processReleaseDate(releaseDate: String): String {
+        return if (releaseDate.isNotBlank()) {
+            val inputFormat = SimpleDateFormat("yyyy-MM-dd")
+            val date = inputFormat.parse(releaseDate)
+            val outputFormat = SimpleDateFormat("MM/dd/yyyy")
+            outputFormat.format(date)
+        } else {
+            return ""
+        }
     }
 }
