+++ app/src/main/java/greenberg/moviedbshell/ViewHolders/PopularMovieViewHolder.kt
@@ -7,25 +7,22 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
-import com.bumptech.glide.Glide
-import com.bumptech.glide.request.RequestOptions
 import greenberg.moviedbshell.Models.PopularMoviesModels.PopularMovieResultsItem
 import greenberg.moviedbshell.MosbyImpl.PopularMoviesPresenter
 import greenberg.moviedbshell.R
-import java.text.SimpleDateFormat
 
 class PopularMovieAdapter(var popularMovieList: MutableList<PopularMovieResultsItem?> = mutableListOf(),
-                          var popularMoviesPresenter: PopularMoviesPresenter) : RecyclerView.Adapter<RecyclerView.ViewHolder>() {
+                          private val presenter: PopularMoviesPresenter) : RecyclerView.Adapter<RecyclerView.ViewHolder>() {
 
     override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
         if (holder is PopularMovieViewHolder) {
             val currentItem = popularMovieList.getOrNull(position)
             //todo: load posters and have like, placeholders
             holder.cardItemTitle.text = currentItem?.title
-            holder.cardItemReleaseDate.text = currentItem?.releaseDate?.let { processReleaseDate(it) }
+            holder.cardItemReleaseDate.text = currentItem?.releaseDate?.let { presenter.processReleaseDate(it) }
             holder.cardItemOverview.text = currentItem?.overview
-            popularMovieList[position]?.let { fetchPoster(holder.cardItemPosterImage, it) }
-            holder.cardItem.setOnClickListener { popularMoviesPresenter.onCardSelected(currentItem?.id ?: -1) }
+            popularMovieList[position]?.let { presenter.fetchPoster(holder.cardItemPosterImage, it) }
+            holder.cardItem.setOnClickListener { presenter.onCardSelected(currentItem?.id ?: -1) }
         }
     }
 
@@ -33,29 +30,6 @@ class PopularMovieAdapter(var popularMovieList: MutableList<PopularMovieResultsI
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
         return PopularMovieViewHolder(LayoutInflater.from(parent.context).inflate(R.layout.popular_movie_card, parent, false))
-    }
-
-    private fun fetchPoster(cardItemPosterView: ImageView, item: PopularMovieResultsItem) {
-        //Load poster art
-        item.posterPath?.let {
-            Glide.with(cardItemPosterView)
-                    .load(cardItemPosterView.context.getString(R.string.poster_url_substitution, it))
-                    .apply { RequestOptions().centerCrop() }
-                    .into(cardItemPosterView)
-        }
-    }
-
-    //TODO: probably make sure every date is like this?
-    //there has to be a better way to do this
-    private fun processReleaseDate(releaseDate: String): String {
-        return if (releaseDate.isNotBlank()) {
-            val inputFormat = SimpleDateFormat("yyyy-MM-dd")
-            val date = inputFormat.parse(releaseDate)
-            val outputFormat = SimpleDateFormat("MM/dd/yyyy")
-            outputFormat.format(date)
-        } else {
-            return ""
-        }
     }
 
     class PopularMovieViewHolder(val view: View) : RecyclerView.ViewHolder(view) {
